requirements.txt
matplotlib
numpy
Pillow
tensorflow
tensorflow-hub
tensorflow-metal
jinja2
pdfkit
